BPE26-ref: extract playback feature to dedicated file, move colors util where appropriate
