+++ src/features/ui/index.ts
@@ -1,4 +1,12 @@
 export { createFullscreenButtonSection } from "./fullscreen-button";
+export {
+  createPlaybackControlPanel,
+  handlePlayPause,
+  handleTimeBackward,
+  handleTimeForward,
+  handleTrackBackward,
+  handleTrackForward,
+} from "./playback";
 export { setupPlayerStickiness } from "./player-stickiness";
 export { createProgressBar, updateProgressBar } from "./progress";
 export { createVolumeControlSection, handleMuteToggle, syncMuteBtn } from "./volume";
